+++ src/DemoApp/tailwind.config.js
@@ -8,6 +8,24 @@ module.exports = {
         "./Components/**/*.{razor,html,cshtml,cs}",
     ],
     darkMode: 'class',
+    safelist: [
+        "md:bg-transparent",
+        "md:block",
+        "md:border-0",
+        "md:dark:hover:bg-transparent",
+        "md:dark:hover:text-white",
+        "md:flex-row",
+        "md:font-medium",
+        "md:hidden",
+        "md:hover:bg-transparent",
+        "md:hover:text-primary-700",
+        "md:mt-0",
+        "md:p-0",
+        "md:space-x-8",
+        "md:text-primary-700",
+        "md:text-sm",
+        "md:w-auto"
+    ],
     theme: {
         extend: {
             colors: {