+++ src/Flowbite.ExtendedIcons/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.0.4-alpha
+
+- Add ScrollTextIcon
+- Add ShapesIcon
+- Add SrcFileIcon
+- Add TestTube2Icon
+- Add CubeIcon  
+
 ## 0.0.2-alpha
 
 - Rolled all assemblies to v0.0.2-alpha